added new app admin dashboard added with funcationality of event admin actions
