+++ dxb/settings.py
@@ -69,6 +69,7 @@ INSTALLED_APPS = [
     "mentor",
     "group",
     "chat",
+    "admindashboard",
     
 ]
