+++ dxb/urls.py
@@ -33,6 +33,7 @@ urlpatterns = [
     path('mentor/', include('mentor.urls')),
     path('group/', include('group.urls')),
     path('chat/', include('chat.urls')),
+    path('admindashboard/', include('admindashboard.urls')),
 ]
 
 if settings.DEBUG:
